Package renamed, floats turned into doubles, new Response Model created
