+++ src/main/java/exercise_screens/ExerciseCreationController.java
@@ -1,4 +1,4 @@
-package exerciseGUI;
+package exercise_screens;
 
 public class ExerciseCreationController {
 }
