+++ src/main/java/exercise_screens/ExerciseCreationScreen.java
@@ -1,4 +1,4 @@
-package exerciseGUI;
+package exercise_screens;
 
 import javax.swing.*;
 import java.awt.*;
